fix: resolve all clippy warnings and errors
- Fix needless_range_loop warnings
- Fix vec_init_then_push warnings
- Fix new_without_default warnings
- Fix redundant_pattern_matching warnings
- Fix single_match warnings
- Fix manual_flatten warnings
- Fix needless_borrows_for_generic_args warnings
- Fix field_reassign_with_default warnings
- Fix suspicious_open_options warnings
- Fix unnecessary_cast warnings
- Fix useless_vec warnings
- Fix ptr_arg warnings

Zero warnings and errors remain across all targets and features.

+++ examples/concurrent_safety_test.rs
@@ -165,7 +165,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
         println!("  ✓ Final counter value: {} (started at 0)", final_value);
 
-        if final_value <= successes as i32 {
+        if final_value <= successes {
             println!("  ✓ Counter consistency maintained");
         } else {
             panic!(
